+++ src/main/java/com/example/firstcrud/repositories/IContratRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IContratRepository extends CrudRepository<Contrat,Integer> {
+public interface IContratRepository extends CrudRepository<Contrat,Long> {
 
 
 }
